refactor(type_checker): extract check_expression into expr.rs
